A CUDA Interface for evolutionary optimization
This was a CUDA Interface I wrote to make evolving the weights of
artificial neural networks much more simpler.  I wrote this code about 2
years ago. These days Theano will be a better option for doing this, but
prior to theano and caffe being mainstream, there were not a lot of
tools for GPU Computing.